Add page folder and Home.jsx
Refactor App.jsx
